+++ Shared/CommonAppDelegate.swift
@@ -22,8 +22,6 @@ import dydxStateManager
 import dydxViews
 import dydxAnalytics
 import StatsigInjections
-import dydxCartera
-import dydxFormatter
 
 open class CommonAppDelegate: ParticlesAppDelegate {
     open var notificationTag: String {
@@ -68,9 +66,6 @@ open class CommonAppDelegate: ParticlesAppDelegate {
         injectAmplitude()
         FeatureService.shared?.activate { /* [weak self] in */
             Injection.shared?.injectFeatured(completion: completion)
-        }
-        if dydxBoolFeatureFlag.privy_ios.isEnabled {
-            injectPrivy()
         }
     }
 
@@ -143,19 +138,6 @@ open class CommonAppDelegate: ParticlesAppDelegate {
             return
         }
         StatsigFeatureFlagsProvider.shared = StatsigFeatureFlagsProvider(apiKey: apiKey, userId: dydxCompositeTracking.getStableId(), environment: environment)
-    }
-    
-    open func injectPrivy() {
-        Console.shared.log("injectPrivy")
-        guard let privyAppId = CredientialConfig.shared.credential(for: "privyAppId") else {
-            assertionFailure("privyAppId is missing")
-            return
-        }
-        guard let privyClientId = CredientialConfig.shared.credential(for: "privyClientId") else {
-            assertionFailure("privyClientId is missing")
-            return
-        }
-        PrivyAuthManager.shared = PrivyAuthManager(appId: privyAppId, appClientId: privyClientId)
     }
 
     open func injectAttribution() {
