+++ dydxV4/dydxV4/AppDelegate.swift
@@ -22,6 +22,7 @@ import UIAppToolkits
 import UIToolkits
 import Utilities
 import dydxAnalytics
+import dydxCartera
 
 #if _iOS
     import FirebaseStaticInjections
@@ -69,6 +70,14 @@ class AppDelegate: CommonAppDelegate {
             if let document = Directory.document {
                 Directory.delete(document)
             }
+        }
+    }
+    
+    override func injectFeatures(completion: @escaping () -> Void) {
+        super.injectFeatures(completion: completion)
+        
+        if dydxBoolFeatureFlag.privy_ios.isEnabled {
+            injectPrivy()
         }
     }
 
@@ -160,5 +169,18 @@ class AppDelegate: CommonAppDelegate {
             Console.shared.log("CarteraConfig SDK throwing error: \(error)")
             return true
         }
+    }
+    
+    open func injectPrivy() {
+        Console.shared.log("injectPrivy")
+        guard let privyAppId = CredientialConfig.shared.credential(for: "privyAppId") else {
+            assertionFailure("privyAppId is missing")
+            return
+        }
+        guard let privyClientId = CredientialConfig.shared.credential(for: "privyClientId") else {
+            assertionFailure("privyClientId is missing")
+            return
+        }
+        PrivyAuthManager.shared = PrivyAuthManager(appId: privyAppId, appClientId: privyClientId)
     }
 }
